requirements.txt
Django==1.6.2
MySQL-python==1.2.3
PIL==1.1.7
South==0.8.4
django-appconf==0.6
django-filebrowser==3.5.4
django-grappelli==2.5.1
django-imagekit==3.2
django-tinymce==1.5.2
django-tinymce-images==0.1.0
pilkit==1.1.6
pillow==2.3.0
six==1.5.2
